requirements.txt
flask
gunicorn
psycopg2-binary
flask-login
werkzeug